Adds input functionality for (W,A,S,D, Space), also allows multiple meshes to be loaded into the vertex and index buffer. Finally adds precision to the GameTime class elapsed time getter

+++ PrototypeTwo/SGE/SGE/GameTime.cpp
@@ -22,5 +22,5 @@ float GameTime::GetDeltaTime(){
 }
 
 float GameTime::GetElapsedTime(){
-	return elapsedTime/CLOCKS_PER_SEC;
+	return (float)elapsedTime/CLOCKS_PER_SEC;
 }
